Fix a typecast warning in glfw clipboard access (#4479)

+++ src/platforms/rcore_desktop_glfw.c
@@ -988,7 +988,7 @@ Image GetClipboardImage(void)
     }
     else
     {
-        image = LoadImageFromMemory(".bmp", fileData, dataSize);
+        image = LoadImageFromMemory(".bmp", fileData, (int)dataSize);
     }
     return image;
 }
